mcp_can: Constructor changes.
The original implementation of mcp_can was expecting a variable
SPIClass SPI to be defined. To avoid such bindings to other libraries
the SPI object is now given at construction.

The original implementation also made assumptions about how to initiate
the CS pin and how to control it. The arduino way doesn't fit all systems,
so in order to make a looser coupling to arduino a chipSelect function
can now be given at construction.
Configuration of the pin is not considered a concern for the mcp_can class.

Except from the new SPI argument, the old constructor is left untouched in
an attempt to make only needed changes. Unfortunately the
MCP_SELECT/UNSELECT macros have been changed so the class isn't backward
compatible.